added MatrixTypes.h and Operations.h

+++ main.cpp
@@ -1,22 +1,22 @@
 #include "Matrix.h"
-
+#include "Operations.h"
 
 int main()
 {
 
-	vector<int> row = {1,2,-3};
-	vector<int> col = {4,-5,6};
-	vector<int> mat = { 1,0,0,
+	vector<double> row = {1,2,-3};
+	vector<double> col = {4,-5,6};
+	vector<double> mat = { 1,0,0,
 					    0,1,0,
 					    0,0,1 };
 
-	Row<int> r(row);
-	Column<int> c(col);
-	Matrix2D<int> M(3,3,mat);
+	Row<double> r(row);
+	Column<double> c(col);
+	Matrix2D<double> M(3,3,mat);
 
-	Matrix2D<int> M2( 5, 5 );
+	Matrix2D<double> M2( 5, 5 );
 
-	Matrix2D<int> res = M2^2;
+	Matrix2D<double> res = M2^2;
  
 	res.show();
